Implemented and tested TransSummaryService

+++ src/Domain/Entities/RateEntity.cs
@@ -9,14 +9,14 @@ namespace Domain.Entities
 
         public string To { get; private set; }
 
-        public double Rate { get; private set; }
+        public decimal Rate { get; private set; }
 
         public RateEntity()
         {
             //Required by EF
         }
 
-        public RateEntity(string from, string to, double rate)
+        public RateEntity(string from, string to, decimal rate)
         {
             From = from ?? throw new ArgumentNullException(nameof(from));
             To = to ?? throw new ArgumentNullException(nameof(to));
